requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.5.0
djangorestframework==3.12.1
mysql-connector-python==8.0.22
mysqlclient @ file:///H:/githubProject/MLWA/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
Pillow==8.1.0
protobuf==3.14.0
PyMySQL==0.10.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
xlwt==1.3.0
